+++ src/Headroom/FileSupport/Haskell.hs
@@ -1,8 +1,10 @@
+{-# LANGUAGE LambdaCase        #-}
 {-# LANGUAGE NoImplicitPrelude #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE QuasiQuotes       #-}
 {-# LANGUAGE RecordWildCards   #-}
 {-# LANGUAGE TupleSections     #-}
+
 
 {-|
 Module      : Headroom.FileSupport.Haskell
@@ -20,8 +22,11 @@ parsed /source code files/ as /template variables/. Such info includes
 -}
 
 module Headroom.FileSupport.Haskell
-  ( extractModuleName
+  ( -- * Variables Extraction
+    extractModuleName
   , extractVariablesHaskell
+    -- * Helper Functions
+  , updateYears
   )
 where
 
@@ -32,7 +37,8 @@ import           Headroom.FileSupport.Haskell.Haddock
                                                 ( HaddockModuleHeader(..)
                                                 , extractModuleHeader
                                                 )
-import           Headroom.Regex                 ( match'
+import           Headroom.Regex                 ( gsub
+                                                , match'
                                                 , re'
                                                 )
 import           Headroom.Types                 ( HeaderConfig(..)
@@ -41,7 +47,9 @@ import           Headroom.Types                 ( HeaderConfig(..)
 import           Headroom.Variables             ( mkVariables )
 import           RIO
 import qualified RIO.List                      as L
+import           RIO.Partial                    ( read )
 import qualified RIO.Text                      as T
+
 
 
 -- | Extracts name of /Haskell/ module from given source code file content.
@@ -84,3 +92,36 @@ extractVariablesHaskell _ headerPos text = (mkVariables . catMaybes)
   HaddockModuleHeader {..} = extractModuleHeader headerText
   headerText               = maybe "" (\(s, e) -> cut s e text) headerPos
   cut s e = T.unlines . L.take (e - s) . L.drop s . T.lines
+
+
+-- | Updates years and years ranges in given text.
+--
+-- >>> updateYears 2020 "Copyright (c) 2020"
+-- "Copyright (c) 2020"
+--
+-- >>> updateYears 2020 "Copyright (c) 2019"
+-- "Copyright (c) 2019-2020"
+--
+-- >>> updateYears 2020 "Copyright (c) 2018-2020"
+-- "Copyright (c) 2018-2020"
+--
+-- >>> updateYears 2020 "Copyright (c) 2018-2019"
+-- "Copyright (c) 2018-2020"
+updateYears :: Integer
+            -- ^ current year
+            -> Text
+            -- ^ text to update
+            -> Text
+            -- ^ text with updated years
+updateYears year = processYear . processRange
+ where
+  processYear  = gsub [re'|(?!\d{4}-)(?<!-)(\d{4})|] processYear'
+  processRange = gsub [re'|(\d{4})-(\d{4})|] processRange'
+  replaceYear curr | read curr == year = show year
+                   | otherwise         = mconcat [curr, "-", show year]
+  replaceRange full fromY toY | read toY == year = full
+                              | otherwise = mconcat [fromY, "-", show year]
+  processYear' _    (curr : _) = replaceYear curr
+  processYear' full _          = full
+  processRange' full (fromY : toY : _) = replaceRange full fromY toY
+  processRange' full _                 = full