+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 ## 0.3.0.0 (not released yet, TBD)
-- [#25] Extract dynamic variables from Haskell source code files
+- [#25] Extract _dynamic variables_ from Haskell source code files
+- [#29] Add _dynamic variable_ for current year and add support for updating years
 - [#30] Allow workflow without configuration and template files
 - [#49] Microsite for documentation ([https://doc.norcane.com/headroom/latest/](https://doc.norcane.com/headroom/latest/))
-- Bump _LTS Haskell_ to `15.13`
+- [#51] Allow template variable values to be templates itself
+- Bump _LTS Haskell_ to `15.14`
 
 ## 0.2.2.1 (released 2020-05-08)
 - Minor improves in generated `.headroom.yaml` in `init` mode